Fix an information disclosure via /proc/self/fd/* links
Previously it was possible to obtain information about the location of
the file in the (outer) filesystem by calling readlink() with the path
of stdin in the proc filesystem (/proc/self/fd/0). Additionally, this
made it possible to obtain the case name from the filename.

This change now adds a flag (-Z) to close all open file descriptors to
ensure we are not leaking any additional information, and now the -0,
-1, -2 flags bind-mount the files into /dev/{stdin,stdout,stderr}
respectively and open those files from within the jail instead.