Improve the way uid/gid changes in unprivileged userns
This change uses whatever was passed into minijail_change_[ug]id() as
the user to change in the user namespace. This is used to fix an issue
where calling open(2) on a file on the tmpfs created by minijail would
return EOVERFLOW[1]. An easy way to reproduce is running this on a 4.8
kernel (or Ubuntu Xenial, which has this change backported):

  $ ./minijail0 -T static -Ut -- /bin/bash -c 'touch /tmp/foo'

This change allows a non-zero uid/gid to be mapped to the current user
when entering a namespace, to avoid the above issue.

1: More information about the bug here:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1659087

Bug: None
Test: make tests
Test: ./minijail0 -T static -Ut -u 1000 -g 1000 -M -m -- \
      /bin/bash -c 'touch /tmp/foo'